fix(nvim): cmake LSP server should use local build area
the default build directory is being used by containerized builds
creating a conflict between environments when the LSP server updates
python library paths that are managed inside the CMakeCache.txt.

To avoid conflicts between the environments, the LSP server is
modified to use the environment that would be setup by cmake based
builds.