fix(nvim): nvim-orgmode plugin no longer depends on tree-sitter
Errors are generated when attempting to load "org" support in
nvim-treesitter as it is no longer supportd. nvim-treesitter has removed
support for `org` files as the repo was not updated.

nvim-orgmode is providing the tree sitter syntax directly and no
longer depends on nvim-treesitter for this support.

Follow the updated orgmode setup instructions.